# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'TelU Cup Backend API'
  description: 'Dokumentasi lengkap API manajemen kompetisi internal Tel-U Cup. Mencakup autentikasi, manajemen cabang olahraga, kontingen, registrasi tim, dan bagan pertandingan gugur.'
  version: 2.0.0
servers:
  -
    url: 'http://localhost:8000'
    description: 'Development Server'
paths:
...
  /api/sports:
    get:
      tags:
        - Sports
      summary: 'Daftar semua cabang olahraga'
      description: 'Mengembalikan semua cabang olahraga beserta sub-kategori dan batas maksimal anggota tim. Dapat diakses publik.'
      operationId: listSports
      responses:
        200:
          description: Berhasil
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    type: array
                    items:
                      properties:
                        id:
                          type: integer
                          example: 1
                        name:
                          type: string
                          example: Badminton
                        icon_path:
                          type: string
                          example: icons/badminton.png
                        max_members:
                          type: integer
                          example: null
                        categories:
                          type: array
                          items:
                            properties:
                              id:
                                type: integer
                                example: 1
                              name:
                                type: string
                                example: 'Tunggal Putra'
                              max_members:
                                type: integer
                                example: 1
                            type: object
                      type: object
                type: object
    post:
      tags:
        - Sports
      summary: 'Buat cabang olahraga baru'
      description: 'Admin/panitia membuat cabang olahraga baru. Jika sport memiliki sub-kategori (misal Badminton Putra/Putri), sertakan array `categories`. Jika tidak, sertakan `max_members` langsung.'
      operationId: createSport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
              properties:
                name:
                  type: string
                  example: Futsal
                icon_path:
                  type: string
                  example: icons/futsal.png
                max_members:
                  description: 'Diisi jika sport TIDAK memiliki sub-kategori'
                  type: integer
                  example: 5
                categories:
                  description: 'Diisi jika sport memiliki sub-kategori'
                  type: array
                  items:
                    properties:
                      name:
                        type: string
                        example: 'Tunggal Putra'
                      max_members:
                        type: integer
                        example: 1
                    type: object
              type: object
      responses:
        201:
          description: 'Cabang olahraga berhasil dibuat'
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: 'Cabang olahraga berhasil dibuat.'
                  data:
                    $ref: '#/components/schemas/SportObject'
                type: object
        422:
          description: 'Validasi gagal'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          bearerAuth: []
  '/api/sports/{id}':
    get:
      tags:
        - Sports
      summary: 'Detail satu cabang olahraga'
      description: 'Mengembalikan detail cabang olahraga termasuk semua sub-kategori dan batas anggota.'
      operationId: showSport
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: Berhasil
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: success
                  data:
                    $ref: '#/components/schemas/SportObject'
                type: object
        404:
          description: 'Tidak ditemukan'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags:
        - Sports
      summary: 'Update cabang olahraga'
      description: 'Admin/panitia mengupdate data cabang olahraga. Jika `categories` dikirim, seluruh sub-kategori lama akan digantikan.'
      operationId: updateSport
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  example: Futsal
                icon_path:
                  type: string
                  example: icons/futsal.png
                max_members:
                  type: integer
                  example: 5
                categories:
                  type: array
                  items:
                    properties:
                      name:
                        type: string
                        example: 'Ganda Putra'
                      max_members:
                        type: integer
                        example: 2
                    type: object
              type: object
      responses:
        200:
          description: 'Berhasil diperbarui'
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: success
                  message:
                    type: string
                    example: 'Cabang olahraga berhasil diperbarui.'
                  data:
                    $ref: '#/components/schemas/SportObject'
                type: object
        404:
          description: 'Tidak ditemukan'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          bearerAuth: []
    delete:
      tags:
        - Sports
      summary: 'Hapus cabang olahraga'
      description: 'Admin/panitia menghapus cabang olahraga. Sub-kategori terhapus otomatis (cascade). Player yang terhubung akan memiliki sport_id = null.'
      operationId: deleteSport
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Berhasil dihapus'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        404:
          description: 'Tidak ditemukan'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        -
          bearerAuth: []
...
components:
  schemas:
    ErrorResponse:
      description: 'Format respon error standar'
      properties:
        status:
          type: string
          example: error
        message:
          type: string
          example: 'Pesan error yang menjelaskan penyebab kegagalan.'
      type: object
    SuccessMessage:
      description: 'Format respon sukses tanpa data'
      properties:
        status:
          type: string
          example: success
        message:
          type: string
          example: 'Operasi berhasil.'
      type: object
...
    SportCategoryObject:
      description: 'Sub-kategori dari sebuah cabang olahraga'
      properties:
        id:
          type: integer
          example: 2
        sport_id:
          type: integer
          example: 1
        name:
          type: string
          example: 'Tunggal Putra'
        max_members:
          type: integer
          example: 1
          nullable: true
        created_at:
          type: string
          format: date-time
      type: object
    SportObject:
      description: 'Cabang olahraga beserta sub-kategorinya'
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Badminton
        icon_path:
          type: string
          example: icons/badminton.png
          nullable: true
        max_members:
          description: 'Diisi jika sport tidak memiliki sub-kategori'
          type: integer
          example: null
          nullable: true
        categories:
          type: array
          items:
            $ref: '#/components/schemas/SportCategoryObject'
        created_at:
          type: string
          format: date-time
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      bearerFormat: JWT
      scheme: bearer
...